+++ interfaces/GeneralTransformer.ts
@@ -1,0 +1,7 @@
+import {Observable} from 'rxjs';
+
+export interface GeneralTransformer {
+
+    process(observable: Observable<any>): Observable<any>;
+
+}